Implement add raw top movies for user_id 0
