+++ pyproject.toml
@@ -8,6 +8,9 @@ packages = []
 
 [tool.poetry.dependencies]
 python = "^3.10"
+flake8 = "^6.0.0"
+pre-commit = "^3.0.4"
+isort = "^5.12.0"
 
 [tool.poetry.group.recom.dependencies]
 pyspark = "3.3.1"
