+++ recommender_system/src/engine/spark.py
@@ -1,4 +1,3 @@
-from abc import ABC, abstractmethod
 from pyspark.sql import SparkSession
