+++ recommender_system/src/utilities/indexer.py
@@ -1,8 +1,7 @@
 from pyspark.ml.feature import StringIndexer
 from pyspark.sql import DataFrame
-
-from src.core.settings import AlsHeadersCol
 from pyspark.sql.functions import col
+from src.core.settings import AlsHeadersCol
 
 
 class Indexer:
@@ -16,6 +15,7 @@ class Indexer:
         items_original = data.select(self.headers_col.item_col).distinct()
         self.users = self._indexing(users_original, self.headers_col.user_col)
         self.items = self._indexing(items_original, self.headers_col.item_col)
+        del users_original, items_original
 
     @classmethod
     def _indexing(cls, data: DataFrame, column_name: str) -> DataFrame:
