+++ setup.cfg
@@ -1,0 +1,7 @@
+[flake8]
+max-line-length = 119
+exclude =
+  .venv
+per-file-ignores =
+  test_*.py: S101,DAR101,D100
+  config.py: W291
